fix: move folder content 1 directory up. Delete former directory
