+++ src/Components/ExercisePage/QueryListElement.tsx
