+++ src/Components/StatisticsPage/PassedStatistics.tsx
