+++ src/Components/StatisticsPage/QueryStatisticsListElement.tsx
