+++ src/Pages/LandingPage.tsx
@@ -1,5 +1,5 @@
 import { ExerciseListElement } from "../Components/LandingPage/ExerciseListElement"
-import {ScrollableList} from "../Components/ScrollableList"
+import {ScrollableList} from "../Components/Shared/ScrollableList"
 export const LandingPage = () => {
     const listView = [<ExerciseListElement exercise="a"/>, <ExerciseListElement exercise="b"/>]
     return (<>
